// src/types/Vexatious.ts
export type Vexatious = {
  age: number;
  salary: number;
  position: string;
  department: string;
  years: string;

  healthStatus: string;
  unfit: string;

  peopleWorked: number;
  employerReason: string;
  otherReason: string;

  eventsExperience:[string],
  otherEvent: string;
  dismissalAffected:[string],
  
  evidence:[string],


  progress: number;
  nextProgress: number;
  checked: boolean;

  user: string;
  caseId: string;
};